Replace former front/context/auth files by new ones
