+++ src/nextjs-front/context/auth/authContext.ts
@@ -1,23 +1,43 @@
 import { createContext } from 'react';
 
-export type AuthContextType = {
-	fetchAsLoggedUser: (input: RequestInfo, init?: RequestInit | undefined) => Promise<Response>;
-	getUserData: () => any;
-	authenticateUser: () => Promise<boolean>;
-	logout: () => void;
-	isPreAuthenticated: boolean;
-	isAuthenticated: boolean;
-	setIsAuthenticated: (data: any) => any;
-	setIsPreAuthenticated: (data: any) => any;
-	setUserData: (data: any) => any;
-	clearUser: () => void
-	mergeUserData: (data: any) => any;
-	token: string;
-	setToken: (data: any) => any;
-	isChatOpened: boolean;
- 	setIsChatOpened: (data: any) => any;
-}
+export type LoginMethod = '42' | 'credentials';
+export type LoginPayload =
+    | {
+          email: string;
+          password: string;
+      }
+    | {
+          apiCode: string;
+      };
 
-const authContext = createContext<AuthContextType | null>(null);
+export type AuthContextValue = {
+    authenticateUser: () => Promise<null | User>;
+    session: UserSession;
+    logout: () => Promise<void>;
+    login: (method: LoginMethod, payload: LoginPayload) => Promise<boolean>;
+    loginWithTfa: (userId: string, tfaCode: string) => Promise<User | null>;
+};
 
-export default authContext;
+type LoginAction = 'tfa_required' | 'login' | 'logout' | 'none';
+
+export type UserSession = {
+    user: User | null;
+    state: AuthState;
+    lastLoginAction: {
+        action: LoginAction;
+        queryString?: string;
+    };
+};
+
+/* TODO: replace user by proper type to enforce type safety */
+export type User = any;
+
+export type AuthState =
+    | 'loading'
+    | 'tfa_required'
+    | 'authenticated'
+    | 'unauthenticated';
+
+const AuthContext = createContext<null | AuthContextValue>(null);
+
+export default AuthContext;
